% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActStd.R
\name{ActStd}
\alias{ActStd}
\title{Action Standard}
\usage{
ActStd(
  dataset,
  type = "npd",
  liking,
  expected = NULL,
  concept = NULL,
  prop_lik = 60,
  prop_exp = 60,
  prop_con = 75,
  anota = NULL,
  prod_cur = NULL,
  prop_pd = 10
)
}
\arguments{
\item{dataset}{The Data to be analyzed}

\item{type}{The type of test: 'npd' or 'sc'}

\item{liking}{The Liking variable}

\item{expected}{The Expected Liking variable}

\item{concept}{The Meeting Expectation variable}

\item{prop_lik}{Acceptable prorpotion for liking}

\item{prop_exp}{Acceptable prorpotion for expected liking}

\item{prop_con}{Acceptable prorpotion for meeting expectation}

\item{anota}{The A-Not A variable}

\item{prod_cur}{The current sample}

\item{prop_pd}{Proportion of Discriminators (for A-Not A)}
}
\value{
The table of Action Standard
}
\description{
Computes the Action Standard in different situations
}
\examples{
NULL
}
