% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActStd_signif.R
\name{ActStd_signif}
\alias{ActStd_signif}
\title{Significance Test for Action Standards}
\usage{
ActStd_signif(dataset, liking, expected, concept)
}
\arguments{
\item{dataset}{The Data to be analyzed}

\item{liking}{The Liking variable}

\item{expected}{The Expected Liking variable}

\item{concept}{The Meeting Expectation variable}
}
\value{
The results of the significance test
}
\description{
Compute the significance test for Action Standards
}
\examples{
NULL
}
