% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANOVAs.R
\name{ANOVAs}
\alias{ANOVAs}
\title{Perform ANOVAs}
\usage{
ANOVAs(dataset, var, test)
}
\arguments{
\item{dataset}{The data to analyze}

\item{var}{Set of variables to analyze}

\item{test}{Post'hoc test: 'HSD' or 'LSD'}
}
\value{
The ANOVA tables and mean tables.
}
\description{
Run 2-way ANOVA and post'hoc test
}
\examples{
NULL
}
