% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarChart.R
\name{BarChart}
\alias{BarChart}
\title{Bar Chart}
\usage{
BarChart(dataset, file, import, exit_q, per_prod = FALSE, plot_adj = "Raw")
}
\arguments{
\item{dataset}{The data to be analyzed}

\item{file}{The file location}

\item{import}{The type of import data}

\item{exit_q}{the variables to represent}

\item{per_prod}{Should the results be split per product}

\item{plot_adj}{Plot adjustments: 'Raw' or 'Transposed'}
}
\value{
The generated graphs
}
\description{
Build the Bart Chart
}
\examples{
NULL
}
