% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CATAAnalysis.R
\name{CATAAnalysis}
\alias{CATAAnalysis}
\title{CATA Analysis for Terms}
\usage{
CATAAnalysis(dataset, var, liking, res_by = "All", prop = 20)
}
\arguments{
\item{dataset}{The Data to analyze}

\item{var}{The CATA variables}

\item{liking}{The Liking variable}

\item{res_by}{'All' or per 'Product'}

\item{prop}{Proportion set to 20}
}
\value{
The results of the CATA Analysis
}
\description{
CATA Analysis for Terms
}
\examples{
NULL
}
