% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CATAComp.R
\name{CATAComp}
\alias{CATAComp}
\title{Analysis of CATA Data}
\usage{
CATAComp(dataset, var1, var2, comp, paircomp = "Yes")
}
\arguments{
\item{dataset}{The data to analyze}

\item{var1}{First set of variables}

\item{var2}{Second set of variables}

\item{comp}{What should be compared: 'Type' or 'Product'}

\item{paircomp}{Include pair comparison ('Yes' or 'No')}
}
\value{
The results of the CATA comparison analysis
}
\description{
Perform analysis on CATA Data by also comparing different sets of CATA list (e.g. product vs. concept)
}
\examples{
NULL
}
