% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Frequencies_plot.R
\name{Frequencies_plot}
\alias{Frequencies_plot}
\title{Frequency Plot}
\usage{
Frequencies_plot(
  res_freq,
  type = "Topbox",
  expectation = NULL,
  topn = "Top3",
  size = 8
)
}
\arguments{
\item{res_freq}{The results from the frequencies function}

\item{type}{The type of plot, raw or topbox}

\item{expectation}{Expectation to control for colouring}

\item{topn}{The topN to consider}

\item{size}{Size of text on axis}
}
\value{
The plot to generate
}
\description{
Plot for frequencies, either based on the raw frequencies or topboxes.
}
\examples{
NULL
}
