% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPScore.R
\name{NPScore}
\alias{NPScore}
\title{Net Promoter Score}
\usage{
NPScore(dataset, var, liking, test = "HSD")
}
\arguments{
\item{dataset}{The data to analuze}

\item{var}{The NPS Score variable}

\item{liking}{The Liking variable}

\item{test}{Significance test: 'HSD' or 'LSD'}
}
\value{
The NPS scores
}
\description{
Compute the NPS Score
}
\examples{
NULL
}
