% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PenaltyAnalysis.R
\name{PenaltyAnalysis}
\alias{PenaltyAnalysis}
\title{Analysis of JAR Data}
\usage{
PenaltyAnalysis(
  dataset,
  var,
  liking,
  prop = 0.2,
  comp = "Product",
  size = 8,
  split = "Yes"
)
}
\arguments{
\item{dataset}{The Data to analyze}

\item{var}{The JAR variables}

\item{liking}{The Liking variable}

\item{prop}{The proportion for the JAR variables}

\item{comp}{Comparison by 'Attribute' or 'Product'}

\item{size}{Control the size of the text on x-axis label}

\item{split}{Split the results by 'Type' ('Yes'/'No')}
}
\value{
The results of the penalty analysis
}
\description{
Analysis of JAR Data including Frequency and Penalty Analysis
}
\examples{
NULL
}
