% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PieChart.R
\name{PieChart}
\alias{PieChart}
\title{Build Pie Chart}
\usage{
PieChart(dataset, file, import, exit_q, per_prod = FALSE, plot_adj = "Pie")
}
\arguments{
\item{dataset}{The Data to be analyzed}

\item{file}{The file to import}

\item{import}{The type of import ('EQ' or 'Other')}

\item{exit_q}{The variable to use}

\item{per_prod}{Are the proportion product related?}

\item{plot_adj}{Different types of graph: 'Pie', 'Center', 'Raw', 'Transpose'}
}
\value{
The generated plots
}
\description{
Create Pie Chart
}
\examples{
NULL
}
