% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Frequencies.R
\name{Frequencies}
\alias{Frequencies}
\title{Frequency Tables}
\usage{
Frequencies(dataset, var, labels = NULL, test = "No", topn = "Top3")
}
\arguments{
\item{dataset}{The data to analyze}

\item{var}{The variable to analyze}

\item{labels}{The labels attached to the levels for that variable}

\item{test}{'yes' or 'no' for significance test}

\item{topn}{Compute topbox (e.g. 'Top3')}
}
\value{
The results for the frequency analysis
}
\description{
Compute the frequency tables with topboxes for a set of variables
}
\examples{
NULL
}
